tests/wpt/web-platform-tests/js-self-profiling/cross-origin-script-cors.sub.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>

  <script src="resources/profile-utils.js"></script>

  <script crossorigin src="https://{{hosts[alt][]}}:{{ports[https][0]}}/js-self-profiling/resources/profiling-script.js"></script>
</head>
<body>
  <script>
    promise_test(async t => {
      const trace = await ProfilingScript.profileBuiltinsInNewTask();

      // Ensure that signal from the external script was gathered.
      assert_equals(trace.resources.length, 1);
      assert_equals(trace.resources[0], 'https://{{hosts[alt][]}}:{{ports[https][0]}}/js-self-profiling/resources/profiling-script.js');
      assert_greater_than(trace.frames.length, 0);
      assert_greater_than(trace.stacks.length, 0);
      assert_greater_than(trace.samples.length, 0);
    }, 'cors cross-origin script execution is not observable');
  </script>
</body>
</html>
